Reorganize clases and add maven module
