+++ pom.xml
@@ -42,6 +42,7 @@
 
     <modules>
         <module>beanmother-core</module>
+        <module>beanmother-grpc-converter</module>
         <module>beanmother-java8-converter</module>
         <module>beanmother-joda-time-converter</module>
         <module>beanmother-guava-converter</module>
